Added automatic entity uploads
